Add Prisma adapter
New `bash-gres/prisma` subpath with a `createPrismaClient`, `setup`, and
`PgFileSystem` mirroring the Drizzle / postgres.js / node-postgres adapters.
Wires Prisma into the shared `TEST_ADAPTERS` matrix so every parity test
runs against it.

Adapter accommodates Prisma-specific raw-query quirks:

- routes SELECT vs DML to `$queryRawUnsafe` / `$executeRawUnsafe` (the
  engine treats them as different operations and won't return rows from
  `$executeRawUnsafe`)
- splits multi-statement DDL because the prepared-statement protocol
  rejects compound commands; a small SQL splitter handles `'`/`"` quotes,
  `--`/`/* */` comments, and tagged dollar quoting
- treats `SELECT pg_advisory_*_lock(...)` as non-returning (the engine
  refuses to deserialize the `void` pseudo-type)
- coerces `BigInt` results to `Number` so callers see the same shape as
  pg / postgres.js
- pulls Postgres SQLSTATE out of `PrismaClientKnownRequestError.meta.code`
  and rebuilds a `SqlError` so the `FsError` mappings keep firing
- requires a 5-char `[0-9A-Z]` SQLSTATE on raw pg errors so user-thrown
  `FsError` / `FsQuotaError` (codes like `ENOSPC`) aren't misclassified

Also adds explicit `::int` casts on `substr` / `substring` offset and
limit params: Prisma binds JS `number` as `bigint`, which has no matching
overload (`substr(text, bigint, bigint)` doesn't exist). The cast is a
no-op for the other adapters and the values are already range-checked
upstream.

+++ lib/core/filesystem.ts
@@ -1553,12 +1553,15 @@ export class PgFileSystem {
       const sqlOffset = (options?.offset ?? 0) + 1;
       const sqlLimit = options?.limit;
 
+      // Prisma's query engine binds JS numbers as `bigint`, but `substr`'s
+      // strict overload list has no `(text, bigint, bigint)` form. Casting
+      // to int (range checked above) keeps every adapter happy.
       const textExpr =
-        sqlLimit !== undefined ? `substr(content, $3, $4)` : `substr(content, $3)`;
+        sqlLimit !== undefined ? `substr(content, $3::int, $4::int)` : `substr(content, $3::int)`;
       const binaryExpr =
         sqlLimit !== undefined
-          ? `substring(binary_data FROM $3 FOR $4)`
-          : `substring(binary_data FROM $3)`;
+          ? `substring(binary_data FROM $3::int FOR $4::int)`
+          : `substring(binary_data FROM $3::int)`;
 
       const params: (string | number | Uint8Array)[] = [
         this.workspaceId,
